+++ src/RsqlParserNet.Linq/RsqlSortDirection.cs
@@ -1,0 +1,17 @@
+namespace RsqlParserNet.Linq;
+
+/// <summary>
+/// Defines the direction for a sort expression.
+/// </summary>
+public enum RsqlSortDirection
+{
+    /// <summary>
+    /// Sorts from the lowest value to the highest value.
+    /// </summary>
+    Ascending,
+
+    /// <summary>
+    /// Sorts from the highest value to the lowest value.
+    /// </summary>
+    Descending
+}